+++ Repository/CLI/Commands/User/SubCommands/Status.cs
@@ -44,7 +44,7 @@ namespace PDRepository.CLI.Commands.User.SubCommands
                     {
                         Common.User user = _client.UserClient.GetUserInfo(LoginName);
 
-                        Output("\r\nUser details:\r\n", ConsoleColor.Yellow);
+                        OutputNewLine("User details:\r\n", ConsoleColor.Yellow);
 
                         using (TableWriter writer = new TableWriter(_console, padding: 2))
                         {
@@ -52,7 +52,7 @@ namespace PDRepository.CLI.Commands.User.SubCommands
                             WriteTableHeader(writer, "Property", "Value", ConsoleColor.Blue, ConsoleColor.Blue);
 
                             WriteRow(writer, "Name", user.FullName);
-                            WriteRow(writer, "Comment", (!string.IsNullOrEmpty(user.Comment) ? user.Comment : "(none)"));
+                            WriteRow(writer, "Comment", !string.IsNullOrEmpty(user.Comment) ? user.Comment : "(none)");
                             WriteRow(writer, "Status", user.Status, valueColor: (user.Status == Common.UserStatusEnum.Active) ? ConsoleColor.DarkGreen : ConsoleColor.DarkRed);
                             WriteRow(writer, "Blocked", user.Blocked);
                             WriteRow(writer, "Disabled", user.Disabled);