+++ Repository/CHANGELOG.md
@@ -1,3 +1,26 @@
+# Version 1.0.2
+
+## New
+
+- Work with documents from the command line (enumerate and check out documents)
+
+## Enhancements
+
+- Fix to folder recursion
+- Extended CheckOutEventArgs with document version
+- Documentation updates/clarifications
+
+# Version 1.0.1
+
+## New
+
+- Reset user password functionality (library and CLI)
+
+## Enhancements
+
+- Documentation updates/clarifications
+- CLI parameter alignment
+
 # Version 1.0.0
 
 Initial release.